% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDEG.R
\name{showDEG}
\alias{showDEG}
\title{Plot meta-analysis differential expression for a gene between responders (Red) and non-responders (Blue).}
\usage{
showDEG(gene, Type = "Pan-cancer")
}
\arguments{
\item{gene}{character; gene symbol or ID to visualize.}

\item{Type}{character; one of \code{"Pan-cancer"}, \code{"Melanoma"}, \code{"Renal"}, or \code{"Urothelial"}.}
}
\value{
Invisibly returns a \pkg{ggplot2} object; also prints the plot.
}
\description{
Visualizes log-fold changes and confidence intervals for a specified gene
across studies or within a chosen tumor type.
}
\examples{
\dontrun{
showDEG(
  gene   = "PDCD1",
  Type   = "Pan-cancer"
)
}
}
